pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.125.0",
    "google-generativeai>=0.8.6",
    "openai-agents>=0.6.4",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
